drm/tegra: sor: Do not setup DPAUX at probe time
Setting up the DPAUX at probe time doesn't work because DPAUX is in the
same power domain as the SOR, and nothing may have turned on that domain
if the SOR driver didn't probe yet. The result is that register accesses
hang. Since the SOR device depends on the DPAUX device it's pretty much
guaranteed that nothing in the Linux kernel will have enabled the power
domain. The reason why this hasn't caused any problems is that the power
partition is enable by default on hardware reset.

However, firmware exists that will aggressively disable power partitions
before passing control to the kernel, at which point this becomes a real
problem. The solution implemented in this patch is to defer programming
of the mux option until the DPAUX pads are used. There is already an API
to do this, so just pass it a parameter that determines in which mode to
setup the pads. This will automatically make sure that no registers are
accessed before the SOR needs the DPAUX and has already enable the power
partition.

This version of the commit has been tweaked to maintain HDMI support, as
there's currently an ugly situation where HDMI drivers depend on DPAUX
drivers to control pad multiplexing for the DDC lines, requiring us to
support an odd case where we "must" support a DPAUX driver even if we
have no displayport. Worse, the SOR driver currently determines whether
a SOR is driving HDMI or DP based on SOR usage of a dpaux driver. To work
around this, we temporarily add a i2c-dpaux property that specifies a
dpaux port to be used /only/ as an I2C DDC bus.